[NCL-4717]: Fix deletion of temporary builds not to hit 5 minute transaction timeout
